Further cleaned up the samples.
